+++ database/views/intraday/intraday_by_exchange.sql
@@ -1,7 +1,7 @@
 DROP MATERIALIZED VIEW IF EXISTS quicksight.intraday_by_exchange;
 CREATE MATERIALIZED VIEW quicksight.intraday_by_exchange AS
 
-SELECT concat(coalesce(c.trading_day, CURRENT_DATE), ' ', coalesce(c.time_of_day, a.time_of_day))::timestamp AS trading_ts,
+SELECT concat(CURRENT_DATE, ' ', coalesce(c.time_of_day, a.time_of_day))::timestamp AS trading_ts,
        coalesce(c.trading_day, CURRENT_DATE) as trading_day,
        coalesce(c.time_of_day, a.time_of_day) as time_of_day,
        coalesce(c.exchange_id, a.exchange_id) as exchange_id,
@@ -31,4 +31,4 @@ GROUP BY c.trading_day,
          c.exchange_id, a.exchange_id,
          c.updated_ts;
 
-CREATE UNIQUE INDEX intraday_by_exchange_idx ON quicksight.intraday_by_exchange ( exchange_id, time_of_day);
+CREATE UNIQUE INDEX intraday_by_exchange_idx ON quicksight.intraday_by_exchange ( exchange_id, trading_ts, time_of_day);